remove warning

+++ ui.cpp
@@ -5655,12 +5655,12 @@ void UI::draw_everything() {
 
                 for (auto it = change; it; it = it->next) {
                     im::BulletText(
-                        "start = %s, oldend = %s, newend = %s, oldlen = %llu, newlen = %llu",
+                        "start = %s, oldend = %s, newend = %s, oldlen = %d, newlen = %d",
                         it->start.str(),
                         it->old_end.str(),
                         it->new_end.str(),
-                        it->old_text.len,
-                        it->new_text.len
+                        (u32)it->old_text.len,
+                        (u32)it->new_text.len
                     );
                 }
             }
